added 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
 
 int main(void)
 {
-	unsigned long int a = 1, b = 2, tmp;
+	unsigned long long int a = 1, b = 2, tmp;
 	int count = 1;
 
 	while (count <= 98)
 	{
 		if (count != 1)
 			printf(", ");
-		printf("%lu", a);
+		printf("%llu", a);
 		tmp = a;
 		a = b;
 		b = a + tmp;
